25.02.11
이용안내>FAQ, 공지사항, Q&A
XSS 방지 필터

+++ src/main/java/com/system/lms/fo/mapper/FaqMapper.java
@@ -10,5 +10,5 @@ public interface FaqMapper {
 
     List<Map<String, Object>> selectFaqByPaging(Map<String, Object> params);
 
-    Long selectFaqCount();
+    Integer selectFaqCount();
 }
